+++ src/common_labsoft/components/datomic.clj
@@ -3,21 +3,28 @@
             [datomic.api :as d]
             [io.pedestal.log :as log]
             [common-labsoft.datomic.schema :as datomic.schema]
-            [common-labsoft.datomic.api :as datomic.query]
             [common-labsoft.protocols.config :as protocols.config]
-            [common-labsoft.protocols.datomic :as protocols.datomic]))
+            [common-labsoft.protocols.datomic :as protocols.datomic]
+            [clojure.string :as string]))
+
+(defn install-meta-schema! [conn]
+  @(d/transact conn (concat datomic.schema/meta-schema
+                            datomic.schema/meta-enums)))
 
 (defn ensure-schemas! [conn {:keys [schemas enums]}]
+  (install-meta-schema! conn)
   (doseq [en enums]
-    @(d/transact conn (datomic.schema/create-enums en)))
+    @(d/transact conn (datomic.schema/create-enums en))
+    (log/info :datomic :enum-installed :enum en))
   (doseq [schema schemas]
-    @(d/transact conn (datomic.schema/create-schema schema))))
+    @(d/transact conn (datomic.schema/create-schema schema))
+    (log/info :datomic :schema-installed :schema schema)))
 
 (defn create-connection! [endpoint settings]
   (try
     (d/create-database endpoint)
     (let [connection (d/connect endpoint)]
-      (ensure-schemas! settings connection)
+      (ensure-schemas! connection settings)
       connection)
     (catch Exception e
       (log/error :component :datomic
@@ -27,18 +34,17 @@
 (defrecord Datomic [config settings conn]
   component/Lifecycle
   (start [this]
-    (if conn
-      this
-      (let [endpoint (protocols.config/get! config :datomic-endpoint)]
-        (println "Creating Datomic database and connection...")
-        (assoc this :endpoint endpoint
-                    :conn (create-connection! endpoint settings)))))
+    (prn "Creating Datomic database and connection...")
+    (let [endpoint (protocols.config/get! config :datomic-endpoint)]
+      (assoc this :endpoint endpoint
+                  :conn (create-connection! endpoint settings))))
 
   (stop [this]
-    (if-not conn
-      this
-      (do (d/release conn)
-          (dissoc this :conn))))
+    (prn "Stopping Datomic....")
+    (when (string/starts-with? (:endpoint this) "datomic:mem:")
+      (prn "Deleting in memory database....")
+      (d/delete-database (:endpoint this)))
+    (dissoc this :conn))
 
   protocols.datomic/Datomic
   (connection [this]