+++ src/common_labsoft/pedestal/interceptors/schema.clj
@@ -5,11 +5,12 @@
   {:name  ::coerce
    :enter (fn [context]
             (if-let [body (-> context :request :json-params)]
-              (assoc-in context [:request :data] (schema/coerce body schema))
+              (do
+                (assoc-in context [:request :data] (schema/coerce body schema)))
               context))})
 
 (def coerce-output
-  {:name ::coerce-output
+  {:name  ::coerce-output
    :leave (fn [context]
             (if-let [schema (-> context :response :schema)]
               (update-in context [:response :body] #(schema/coerce % schema))