+++ src/common_labsoft/schema.clj
@@ -1,21 +1,37 @@
 (ns common-labsoft.schema
   (:require [schema.coerce :as coerce]
             [schema-tools.core :as schema-tools]
-            [schema.core :as s]))
+            [schema.core :as s]
+            [common-labsoft.time :as time])
+  (:import (schema.core EnumSchema)))
+
+(def time-matchers {time/LocalDate     time/coerce-to-local-date
+                    time/LocalDateTime time/coerce-to-local-date-time})
+
+(def internalize-matchers (coerce/first-matcher [time-matchers
+                                                 coerce/json-coercion-matcher]))
 
 (defn coerce [value schema]
-  (schema-tools/select-schema value schema coerce/json-coercion-matcher))
+  (schema-tools/select-schema value schema internalize-matchers))
+
+(defn coerce-if [value schema]
+  (if (and schema value)
+    (coerce value schema)
+    value))
 
 (declare skel->schema)
-(defn render-schema [schema]
-  (if (map? schema)
-    (skel->schema schema)
-    schema))
+(defn render-schema [meta]
+  (let [schema (or (:schema meta) meta)]
+    (cond
+      (instance? EnumSchema schema) schema
+      (and (not (record? schema)) (map? schema)) (skel->schema schema)
+      (seq? schema) (map render-schema schema)
+      :else schema)))
 
 (defn attribute->schema [[k meta]]
   (if (:required meta)
-    [(s/required-key k) (render-schema (:schema meta))]
-    [(s/optional-key k) (render-schema (:schema meta))]))
+    [(s/required-key k) (render-schema meta)]
+    [(s/optional-key k) (render-schema meta)]))
 
 (defn skel->schema [skeleton]
   (into {} (map attribute->schema skeleton)))