+++ src/data_readers.clj
@@ -1,1 +1,3 @@
-{debug/print common-labsoft.misc/debug}
+{misc/debug common-labsoft.misc/debug
+ new/time   common-labsoft.time/coerce-to-local-date-time
+ new/date   common-labsoft.time/coerce-to-local-date}